backend/requirements.txt
fastapi
uvicorn
requests
python-dotenv
pydantic
beautifulsoup4
chromadb   # VectorDB

numpy<2
torch==2.2.2   # sentence-transformers 依賴
transformers==4.41.2
sentence-transformers==2.7.0  # embedding model
accelerate==0.30.1